\name{pc9_3um_erlotinib}
\alias{pc9_3um_erlotinib}
\title{PC9 cancer cell observations with 3 micro-molar erlotinib at time 0.}
\description{
  PC9 cancer cell observations with 3um erlotinib applied at time 0. Experiment was performed on 2011/9/9 (F07) in the Vito Quaranta laboratory at Vanderbilt University Cancer Biology Center by Darren Tyson. Cells were tracked by nuclear labeling with histone H2B and imaged on a BD Pathway 855 for several days. All numerical values are in hours. Funding was provided by the National Cancer Institute (NCI).
  
This data set was specifically chosen to give the emg.mle function something difficult to work on.
}
\usage{
  data(pc9_3um_erlotinib)
}
%- maybe also 'usage' for other objects documented here.

\value{
   A data frame of lifespan PC9 observations.
}
\author{ Darren Tyson, Shawn Garbett }

\examples{
  data(pc9_3um_erlotinib)
}