% Generated by roxygen2 (4.0.0): do not edit by hand
\name{power.spectrum}
\alias{power.spectrum}
\title{Power Spectrum}
\usage{
power.spectrum(y, samp.freq = 11025, plot = TRUE, plot_range = NULL)
}
\arguments{
\item{y}{An object of class \code{"EMG"} or \code{"Wave"}.}

\item{samp.freq}{The recording frequency. Defaults to 11025 Hz.}

\item{plot}{logical. Should the power spectrum be plotted?
Defaults to \code{TRUE}.}

\item{plot_range}{numeric vector. The range for plotting. Defaults
  to 0 to samp.freq / 2.}
}
\value{
A \code{data.frame} containing the \code{intensity} at
  each \code{frequency} from 0 to \code{samp.freq / 2}.
}
\description{
Calculate and plot power spectrum for a wave.
}
\examples{
pulse <- read.EMG(system.file("extdata",
                              "pulse_1.wav",
                              package = "EMG"))
power.spectrum(pulse)

no_signal <- read.EMG(system.file("extdata",
                                  "amp_on_baseline.wav",
                                  package = "EMG"))
power.spectrum(no_signal)
}
\author{
Kevin Middleton (middletonk@missouri.edu)
}

