% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{pulse_1}
\alias{pulse_1}
\title{Submaximal Contractions}
\format{A \code{\link{Wave-class}} object of EMG (audio) data
  recorded with the amplifier on and surface electrodes connected
  to the biceps brachii.
\tabular{ll}{ Number of Samples: \tab 94208\cr
              Duration (seconds): \tab 2.14\cr
              Samplingrate (Hertz): \tab 44100\cr
              Channels (Mono/Stereo): \tab Mono\cr
              PCM (integer format): \tab TRUE\cr
              Bit (8/16/24/32/64): \tab 16}}
\source{
Recorded by Kevin Middleton.
}
\description{
File recorded with the Backyard Brains EMG SpikerBox on with
surface elecrodes on the biceps brachii with a series of
submaximal contractions.
}
\examples{
pulse_1
plot(pulse_1)
}
\references{
Backyard Brains \url{https://backyardbrains.com/}
}
\keyword{datasets}

