% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{EMG-package}
\alias{EMG}
\alias{EMG-package}
\title{EMG Analysis in R}
\description{
The \code{EMG} package contains functions for reading, plotting,
and analyzing EMG data collected as \code{wav} files such as are
recorded by Backyard Brains \url{https://backyardbrains.com/} EMG
SpikerBox and the iPhone/Android/Adobe Air app.
}
\details{
Several example \code{wav} files are included in the
\code{extdata} directory.

\tabular{ll}{ Package: \tab EMG\cr
              Type: \tab Package\cr
              Version: \tab 0.0.5\cr
              Date: \tab 2014-05-09\cr
              License: \tab GPL-2\cr
              LazyLoad: \tab yes\cr
              LazyData: \tab yes\cr }
}
\author{
Kevin M. Middleton (\url{middletonk@missouri.edu})
}
\references{
Backyard Brains \url{https://backyardbrains.com/}
}
\keyword{package}

